requirements.txt
# Author: Eray Kocabozdoğan
# Student ID: 280201055

# Core ML/NLP libraries
transformers
sentence-transformers
torch
numpy

# Retrieval & Search
faiss-cpu
rank-bm25

# Privacy & Anonymization
presidio-analyzer
presidio-anonymizer
spacy
faker

# Data handling
pandas
datasets

# Utilities
scikit-learn
nltk
jinja2
